Add teacher class
